+++ src/partials/Popover.svelte
@@ -1,0 +1,62 @@
+<script lang="ts">
+  import 'tippy.js/dist/tippy.css'
+  import 'tippy.js/animations/shift-away.css'
+  import tippy from 'tippy.js'
+  import {onMount} from 'svelte'
+
+  let trigger
+  let tooltip
+  let instance
+
+  onMount(() => {
+    instance = tippy(trigger, {
+      appendTo: () => document.body,
+      allowHTML: true,
+      interactive: true,
+      trigger: 'click',
+      animation: 'shift-away',
+      onShow: () => {
+        const [tooltipContents] = tooltip.children
+
+        instance.popper.querySelector('.tippy-content').appendChild(tooltipContents)
+        instance.popper.addEventListener('mouseleave', e => instance.hide())
+        instance.popper.addEventListener('click', e => {
+          if (e.target.closest('.tippy-close')) {
+            instance.hide()
+          }
+        })
+      },
+      onHidden: () => {
+        const [tooltipContents] = instance.popper.querySelector('.tippy-content').children
+
+        tooltip.appendChild(tooltipContents)
+      },
+    })
+
+    return () => {
+      instance.destroy()
+    }
+  })
+</script>
+
+<svelte:window
+  on:scroll={e => {
+    instance.hide()
+  }} />
+
+<svelte:body
+  on:keydown={e => {
+    if (e.key === 'Escape') {
+      instance.hide()
+    }
+  }} />
+
+<div bind:this={trigger}>
+  <slot name="trigger" />
+</div>
+
+<div bind:this={tooltip} class="hidden">
+  <div>
+    <slot name="tooltip" />
+  </div>
+</div>