+++ src/views/person/PersonDetail.svelte
@@ -36,8 +36,33 @@
   let person = database.getPersonWithFallback(pubkey)
   let loading = true
   let showActions = false
+  let actions = []
 
   $: following = $petnamePubkeys.includes(pubkey)
+
+  $: {
+    actions = []
+
+    if (showActions) {
+      actions.push({onClick: share, label: 'Share', icon: 'share-nodes'})
+
+      if (following) {
+        actions.push({onClick: unfollow, label: 'Unfollow', icon: 'user-minus'})
+      } else if (user.getPubkey() !== pubkey) {
+        actions.push({onClick: follow, label: 'Follow', icon: 'user-plus'})
+      }
+
+      if ($canPublish) {
+        actions.push({href: `/messages/${npub}`, label: 'Message', icon: 'envelope'})
+        actions.push({onClick: openAdvanced, label: 'Advanced', icon: 'sliders'})
+      }
+
+      if (user.getPubkey() === pubkey) {
+        actions.push({href: '/profile', label: 'Edit', icon: 'edit'})
+      }
+    }
+  }
+
 
   onMount(async () => {
     log('Person', npub, person)
@@ -73,12 +98,6 @@
 
   const toggleActions = () => {
     showActions = !showActions
-  }
-
-  const makeGetTransition = () => {
-    let i = 0
-
-    return () => ({y: 20, delay: i++ * 30})
   }
 
   const setActiveTab = tab => navigate(routes.person(pubkey, tab))
@@ -142,78 +161,27 @@
         </div>
         <div class="whitespace-nowrap flex gap-3 flex-wrap relative">
           <div on:click|stopPropagation={toggleActions} class="px-5 py-2 cursor-pointer">
-            <i class="fa fa-ellipsis-vertical" />
+            <i class="fa fa-xl fa-ellipsis-vertical" />
           </div>
-          {#if showActions}
-          {@const getTransition = makeGetTransition()}
           <div class="absolute top-0 right-0 mt-12 flex flex-col gap-2 opacity-90">
             <div
               class="absolute inset-0 bg-black rounded-full"
+              class:hidden={!showActions}
               style="filter: blur(15px)"
-              transition:fade />
+              transition:fade|local />
+            {#each actions as {onClick, href, label, icon}, i}
             <div
               class="flex gap-2 justify-end items-center z-10 cursor-pointer"
-              transition:fly={getTransition()}
-              on:click={share}>
-              <div class="text-light">Share</div>
+              in:fly|local={{y: 20, delay: i * 30}}
+              out:fly|local={{y: 20, delay: (actions.length - i - 1) * 30}}
+              on:click={onClick}>
+              <div class="text-light">{label}</div>
               <Anchor type="button-circle">
-                <i class="fa fa-share-nodes" />
+                <i class={`fa fa-${icon}`} />
               </Anchor>
             </div>
-            {#if following}
-            <div
-              class="flex gap-2 justify-end items-center z-10 cursor-pointer"
-              transition:fly={getTransition()}
-              on:click={unfollow}>
-              <div class="text-light">Unfollow</div>
-              <Anchor type="button-circle">
-                <i class="fa fa-user-minus" />
-              </Anchor>
-            </div>
-            {:else if user.getPubkey() !== pubkey}
-            <div
-              class="flex gap-2 justify-end items-center z-10 cursor-pointer"
-              transition:fly={getTransition()}
-              on:click={follow}>
-              <div class="text-light">Follow</div>
-              <Anchor type="button-circle">
-                <i class="fa fa-user-plus" />
-              </Anchor>
-            </div>
-            {/if}
-            {#if $canPublish}
-            <div
-              class="flex gap-2 justify-end items-center z-10 cursor-pointer"
-              transition:fly={getTransition()}
-              on:click={() => navigate(`/messages/${npub}`)}>
-              <div class="text-light">Message</div>
-              <Anchor type="button-circle">
-                <i class="fa fa-envelope" />
-              </Anchor>
-            </div>
-            <div
-              class="flex gap-2 justify-end items-center z-10 cursor-pointer"
-              transition:fly={getTransition()}
-              on:click={openAdvanced}>
-              <div class="text-light">Advanced</div>
-              <Anchor type="button-circle">
-                <i class="fa fa-sliders" />
-              </Anchor>
-            </div>
-            {/if}
-            {#if user.getPubkey() === pubkey}
-            <div
-              class="flex gap-2 justify-end items-center z-10 cursor-pointer"
-              transition:fly={getTransition()}
-              on:click={() => navigate("/profile")}>
-              <div class="text-light">Edit</div>
-              <Anchor type="button-circle">
-                <i class="fa fa-edit" />
-              </Anchor>
-            </div>
-            {/if}
+            {/each}
           </div>
-          {/if}
         </div>
       </div>
       <p>{@html renderContent(person?.kind0?.about || '')}</p>